requirements.txt
numpy
sounddevice
websockets
python-dotenv
azure-identity
azure-cognitiveservices-speech
openai
loguru
langchain
langchain-community
langchain-core
langchain-ollama
langchain-openai
langchain-anthropic
ollama
pyttsx3
pytz
SpeechRecognition
PyAudio
ipython

# Additional tool dependencies
mss
duckduckgo-search
pytesseract

langgraph
chromadb
langchain_chroma

# Gmail dependencies
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Test and quality tooling
pytest
ruff
pre-commit
